requirements.txt
matplotlib
scikit-learn
scipy
tflearn
pandas
protobuf
tensorflow
numpy-mkl
h5py
curses
nltk